Move testing strategy module
