+++ tests/common/strategies.py
