+++ tests/property/test_uint.py
@@ -7,7 +7,7 @@ from eth_abi import (
     decode_single,
 )
 
-from .abi_type_strategies import (
+from ..common.strategies import (
     uint_strs,
     uint_values,
 )
